# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ficha Digital API",
        "description": "API para el sistema de gestión de solicitudes y fichas digitales",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/personas/buscar/{dni_cuil}": {
            "get": {
                "tags": [
                    "Personas"
                ],
                "summary": "Buscar persona por DNI o CUIL",
                "description": "Busca una persona en el padrón externo por DNI o CUIL",
                "operationId": "37b31f56b9e2c8991056cdadea688fc0",
                "parameters": [
                    {
                        "name": "dni_cuil",
                        "in": "path",
                        "description": "DNI o CUIL de la persona (solo números, sin guiones ni puntos)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "20123456789"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Búsqueda exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "estado": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "mensaje": {
                                            "type": "string",
                                            "example": "Persona encontrada"
                                        },
                                        "personas": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error del servidor externo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "No se pudo conectar con el servicio externo"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}